requirements.txt
psycopg[binary]
boto3
requests
python-dotenv
xmltodict
